requirements.txt
CherryPy==18.3.0
Click==7.0
Flask-Cors==3.0.8
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.1
Flask==1.1.1
GetOldTweets3==0.0.11
Jinja2==2.10.3
MarkupSafe==1.1.1
Pattern==3.6
PySocks==1.7.1
SQLAlchemy-serializer==1.3.4.2
SQLAlchemy==1.3.12
Werkzeug==0.16.0
astroid==2.3.3
backports.csv==1.0.7
beautifulsoup4==4.8.1
blis==0.4.1
certifi==2019.11.28
chardet==3.0.4
cheroot==8.2.1
colorama==0.4.3
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.3
dnspython==1.16.0
et-xmlfile==1.0.1
feedparser==5.2.1
future==0.18.1
graphviz==0.8.4
gunicorn==20.0.4
idna==2.6
importlib-metadata==0.23
isort==4.3.21
itsdangerous==1.1.0
jaraco.functools==2.0
jdcal==1.4.1
joblib==0.14.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.4.1
matplotlib==3.1.1
mccabe==0.6.1
more-itertools==7.2.0
murmurhash==1.0.2
mxnet==1.5.0
mysqlclient==1.4.4
nltk==3.4.5
numpy==1.16.6
oauthlib==3.1.0
openpyxl==3.0.3
pandas==0.25.2
pdfminer.six==20191020
plac==1.1.3
portend==2.5
preshed==3.0.2
psutil==5.6.7
pycorenlp==0.3.0
pycryptodome==3.9.0
pylint==2.4.4
pymongo==3.10.0
pyparsing==2.4.5
pyquery==1.4.1
pyspellchecker==0.5.2
python-dateutil==2.8.0
python-docx==0.8.10
python-dotenv==0.10.3
pytz==2019.3
# pywin32==225
requests-oauthlib==1.3.0
# requests==2.18.4
requests==2.20.0
scikit-learn==0.22
scipy==1.4.1
seaborn==0.9.0
six==1.13.0
sklearn==0.0
sortedcontainers==2.1.0
soupsieve==1.9.4
spacy==2.2.2
srsly==0.2.0
stanfordcorenlp==3.9.1.1
tempora==1.14.1
thinc==7.3.1
tqdm==4.38.0
tweepy==3.8.0
typed-ast==1.4.0
urllib3==1.22
wasabi==0.4.0
wrapt==1.11.2
zc.lockfile==2.0
zipp==0.6.0
# en-core-web-lg==2.2.5
# https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-2.2.5/en_core_web_lg-2.2.5.tar.gz
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.5/en_core_web_sm-2.2.5.tar.gz